+++ test/cli/test_versions.py
@@ -1,9 +1,12 @@
+# Copyright (c) 2018-2024 by xcube team and contributors
+# Permissions are hereby granted under the terms of the MIT License:
+# https://opensource.org/licenses/MIT.
+
 from test.cli.helpers import CliTest
 from xcube.util.temp import new_temp_file
 
 
 class VersionsTest(CliTest):
-
     def test_help_option(self):
         result = self.invoke_cli(["versions", "--help"])
         self.assertEqual(0, result.exit_code)