+++ test/core/resampling/test_temporal.py
@@ -1,3 +1,7 @@
+# Copyright (c) 2018-2024 by xcube team and contributors
+# Permissions are hereby granted under the terms of the MIT License:
+# https://opensource.org/licenses/MIT.
+
 import unittest
 
 import numpy as np
@@ -10,7 +14,6 @@ from xcube.core.schema import CubeSchema
 
 
 class ResampleInTimeTest(unittest.TestCase):
-
     def setUp(self) -> None:
         num_times = 30