+++ test/webapi/ows/coverages/test_util.py
@@ -1,10 +1,13 @@
+# Copyright (c) 2018-2024 by xcube team and contributors
+# Permissions are hereby granted under the terms of the MIT License:
+# https://opensource.org/licenses/MIT.
+
 import unittest
 import xcube.core.new
 import xcube.webapi.ows.coverages.util as util
 
 
 class UtilTest(unittest.TestCase):
-
     def setUp(self):
         self.ds_latlon = xcube.core.new.new_cube()
         self.ds_xy = xcube.core.new.new_cube(x_name="x", y_name="y")