+++ xcube/core/extract.py
@@ -1,3 +1,7 @@
+# Copyright (c) 2018-2024 by xcube team and contributors
+# Permissions are hereby granted under the terms of the MIT License:
+# https://opensource.org/licenses/MIT.
+
 from typing import Dict, Union, Tuple, Any, Optional, Mapping, Sequence, Hashable
 
 import dask.array as da