+++ CHANGES.md
@@ -2,6 +2,12 @@
 
 * Embedded [xcube-viewer 1.1.1](https://github.com/xcube-dev/xcube-viewer/releases/tag/v1.1.1).
 
+* Updated copyright notices in all source code files. 
+
+* Fixed xcube plugin auto-recognition in case a plugin project
+  uses `pyproject.toml` file. (#963)
+
+
 ## Changes in 1.5.0
 
 * Enhanced spatial resampling in module `xcube.core.resampling` (#955):